Return focus to opening button when closing modal dialog
This previously worked for the close button but not the Save button
because use of the latter generated a new table in which the calling
button was not a part. This has been fixed by referencing the id of the
relevant calling button (that is the same in the new table) rather than
the html element.